combining settings with its composable
